+++ tests/testthat/test_space_filling.R
@@ -1,5 +1,3 @@
-library(testthat)
-library(dials)
 
 context("space filling designs")